Dispatch 26 compiler emits into oblivion.  Vetting is desired, please   post to the list if you participate.  They are all blindingly obvious,   but extra eyes always help
  This eliminates all but the regex emits and MSVC's borked misdeclaration
  of FD_SET.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@92949 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_config.c
@@ -1566,7 +1566,9 @@ static void default_listeners(pool *p, server_rec *s)
     new = ap_pcalloc(p, sizeof(listen_rec));
     new->local_addr.sin_family = AF_INET;
     new->local_addr.sin_addr = ap_bind_address;
-    new->local_addr.sin_port = htons(s->port ? s->port : DEFAULT_HTTP_PORT);
+    /* Buck ugly cast to get around terniary op bug in some (MS) compilers */
+    new->local_addr.sin_port = htons((unsigned short)(s->port ? s->port 
+                                                        : DEFAULT_HTTP_PORT));
     new->fd = -1;
     new->used = 0;
     new->next = NULL;
